Updates downloaders and fixes errors
Updates both downloaders to make a bigger petition instead of make one by one.
Now the users can create a corpus with its repositories files.
Now when an user or admin logout it will unregister its listeners
(if the user is an user, not an admin, it will delete its corpus temporary folder).